modify according to comments + add effet component
