+++ front/src/modules/spreadsheet-import/components/ModalCloseButton.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 import { useSpreadsheetImportInitialStep } from '@/spreadsheet-import/hooks/useSpreadsheetImportInitialStep';
 import { useSpreadsheetImportInternal } from '@/spreadsheet-import/hooks/useSpreadsheetImportInternal';
 import { IconX } from '@/ui/display/icon/index';
-import { useDialog } from '@/ui/feedback/dialog//hooks/useDialog';
+import { useDialogManager } from '@/ui/feedback/dialog-manager/hooks/useDialogManager';
 import { IconButton } from '@/ui/input/button/components/IconButton';
 import { useStepBar } from '@/ui/navigation/step-bar/hooks/useStepBar';
 
@@ -33,7 +33,7 @@ export const ModalCloseButton = ({ onClose }: ModalCloseButtonProps) => {
     initialStep,
   });
 
-  const { enqueueDialog } = useDialog();
+  const { enqueueDialog } = useDialogManager();
 
   const handleClose = () => {
     if (activeStep === -1) {
