+++ front/src/modules/ui/feedback/dialog-manager/components/Dialog.tsx
