+++ front/src/modules/ui/feedback/dialog-manager/scopes/DialogManagerScope.tsx
