+++ front/src/modules/ui/feedback/dialog-manager/states/dialogInternalScopedState.ts
