+++ front/src/modules/ui/feedback/dialog-manager/types/DialogHotkeyScope.ts
