+++ front/src/modules/ui/feedback/dialog-manager/types/DialogOptions.ts
